+++ include/navkit/core/config/Types.hpp
@@ -1,0 +1,12 @@
+// Copyright (c) 2026 William Gordon Carter.
+// All Rights Reserved.
+
+#pragma once
+
+namespace navkit::core
+{
+
+using Scalar_t = double;
+using Time_t = double;
+
+} // namespace navkit::core